refactor: Convert calendar view into a component
* Moved all calendar view related code from Home.js to CalendarView.js
* Imported CalendarView into Home.js
* Home.js now uses <CalendarView> component call
