+++ front-end/src/pages/Home.js
@@ -1,5 +1,5 @@
 import Container from "react-bootstrap/Container";
-import { useEffect, useState, useRef } from "react";
+import { useState, useRef } from "react";
 import Row from "react-bootstrap/Row";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
@@ -7,8 +7,8 @@ import Col from "react-bootstrap/Col";
 import Card from "react-bootstrap/Card";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
-import Calendar from "react-calendar";
 import TaskList from "../components/TaskList.js"
+import CalendarView from "../components/CalendarView.js"
 
 const Home = () => {
   const [taskName, setNameInput] = useState(""); // New state for the name input
@@ -62,14 +62,6 @@ const Home = () => {
   const handleDescInput = (e) => {
     setDescInput(e.target.value);
   };
-  const [date, setDate] = useState(new Date());
-  const [tasks, setTasks] = useState([]); // this is where we can store the tasks from the database
-
-  useEffect(() => {
-    fetch("https://planned-out-backend-jdx6.onrender.com/")
-      .then((res) => res.json())
-      .then((data) => setTasks(data));
-  }, []);
 
   return (
     <div className="App">
@@ -88,7 +80,6 @@ const Home = () => {
             </Card>
             <Card className="tasks-add">
               <Card.Title>New Task</Card.Title>
-
               <Form.Control
                 ref={inputBox}
                 className="taskNameBox"
@@ -129,28 +120,7 @@ const Home = () => {
             <Card className="react-calendar">
               <Card.Title>Calendar</Card.Title>
               <div className="calendar-container">
-                <Calendar
-                  onChange={setDate}
-                  value={date}
-                  calendarType="gregory"
-                  onClickDay={(value, event) => console.log(value)}
-                  tileContent={({ date }) => {
-                    if (tasks) {
-                      // Convert the date to a string in "MM-DD-YYYY" format
-                      const formattedDate = `${date.getMonth() + 1
-                        }-${date.getDate()}-${date.getFullYear()}`;
-
-                      // Check if there is a task for the selected date
-                      const hasTask = tasks.some(
-                        (task) => task.taskDate === formattedDate
-                      );
-
-                      // Render a custom content if there is a task, otherwise return null
-                      return hasTask ? " 📃" : null;
-                    }
-                    return null;
-                  }}
-                />
+                <CalendarView/>
               </div>
             </Card>
           </Col>
